+++ Totoro/scheduler/plugger.py
@@ -484,10 +484,10 @@ class Plugger(object):
         forcePlugPriority = int(config['plugger']['forcePlugPriority'])
 
         if len(plates) > len(self.carts):
-            log.warning('PLUGGER: {0} plates to allocate but only {1} carts '
-                        'available. Using the first {1} plates.'.format(len(plates),
-                                                                        len(self.carts)),
-                        TotoroPluggerWarning)
+            log.petunia_warning('PLUGGER: {0} plates to allocate but only {1} carts '
+                                'available. Using the first {1} plates.'.format(len(plates),
+                                                                                len(self.carts)),
+                                TotoroPluggerWarning)
             plates = plates[0:len(self.carts)]
 
         # Gets the active pluggings