+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 - For MaNGa time, get only plates with `MaNGA dither` or `MaNGA 10min`. This prevents selecting plugged plates with `MaNGA Globular` survey mode.
 - Output warning about more plates than carts in plugger as `petunia_warning`.
+- Changed default autoscheduler schedule file to ``Sch_base.18_eb_MaStar.txt``.
 
 ### Fixed
 - Fix check of incomplete S/N in a new exposure.